refactor(web/sidebar): introduce new sidebar components for enhanced navigation
- Created the AppSidebar component to serve as the main sidebar, integrating user navigation and chat functionalities.
- Added NavChats and NavChatHistory components to manage chat creation and history display within the sidebar.
- Implemented NavUser component for user profile management and logout functionality.
- Updated import paths for sidebar components to reflect the new structure, improving modularity and organization.
