+++ apps/web/components/sidebar/app-sidebar.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react";
 
-import { NavUser } from "@/components/sidebar-nav/nav-user";
+import { NavUser } from "@/components/sidebar/nav-user";
 import { PicoLogo } from "@/components/pico-logo";
 import {
   Sidebar,
@@ -12,8 +12,8 @@ import {
   SidebarRail,
 } from "@/components/ui/sidebar";
 import { useAuth } from "@/providers/auth-provider";
-import { NavChats } from "@/components/sidebar-nav/nav-chats";
-import { NavChatHistory } from "@/components/sidebar-nav/nav-chat-history";
+import { NavChats } from "@/components/sidebar/nav-chats";
+import { NavChatHistory } from "@/components/sidebar/nav-chat-history";
 
 export function AppSidebar({ ...props }: React.ComponentProps<typeof Sidebar>) {
   const { user } = useAuth();
